Fix bug in UpdateRunLoop that caused CPU Spinning
There is a special case in that if the [NSRunLoop runUntilDate:] doesn't
have any events to process, then it will return nil immediately
without sleeping for the full 60 seconds.  This was causing wasted
cycles in the UpdateRunLoop.  I replaced this call with the contents of
the function in the gnustep source.  Now we use [NSRunLoop runMode: beforeDate:]
which will return NO if there are no events to process.  In that case, then
I can sleep for a period of time to avoid wasting cycles.